\name{experimentSummary}
\alias{experimentSummary}
\title{get summary tables from a experiment}
\usage{
experimentSummary(experiment)
}
\arguments{
  \item{experiment}{a list containing 'frapPair' or 'frap'
  objects, as returned by \code{\link{analyzeFrapPair}} or
  \code{\link{analyzeFrap}}}
}
\value{
a object of class 'experimentSummary' with elements PL,
param and param1
}
\description{
get summary tables from a experiment
}

