\docType{data}
\name{FRAP.options}
\alias{FRAP.options}
\title{List of options for the FRAP experiment analysis}
\format{\preformatted{List of 28
 $ VERSION.STR            : chr "31"
 $ MODEL.PARAMETERS       : chr [1:19] "log10_kEVfrac" "log10_kI" "tot0" "PBn" ...
 $ FIXED.PARAMETERS       : Named num [1:19] 0 0 0 0 0 0 0 1 1 0 ...
  ..- attr(*, "names")= chr [1:19] "log10_kEVfrac" "log10_kI" "tot0" "PBn" ...
 $ VALUE.PARAMETERS       : Named num [1:19] NA NA NA NA 0 4 NA 1 0 0 ...
  ..- attr(*, "names")= chr [1:19] "log10_kEVfrac" "log10_kI" "tot0" "PBn" ...
 $ LOWER.PARAMETERS       : Named num [1:19] NA NA NA 0 0 0 0 0 0 -0.1 ...
  ..- attr(*, "names")= chr [1:19] "log10_kEVfrac" "log10_kI" "tot0" "PBn" ...
 $ UPPER.PARAMETERS       : Named num [1:19] NA NA NA 1 1 8 NA 1 NA 0.1 ...
  ..- attr(*, "names")= chr [1:19] "log10_kEVfrac" "log10_kI" "tot0" "PBn" ...
 $ PRE.ACTIVATION.FRAMES  : num 5
 $ TOTAL.FRAMES.PER.FRAP  : num 35
 $ AUTO.F                 : num 0
 $ F.COR.CYT              : num 1
 $ SEC.PER.FRAME          : num 0.22
 $ PB.CORRECTION.FUN      :function (fl, time)  
 $ SIGMA.FUN              :function (fl)  
 $ SIGMA.METHOD           : chr "SINGLE.CELL.PAIR"
 $ FRAP.ORDER.SEP         : chr "-"
 $ FRAP.ORDER.DAUGHTER.STR: chr "hija"
 $ FRAP.ORDER.MOTHER.STR  : chr "madre"
 $ MOTHER.FRAP.NUMBER     : num 4
 $ DAUGHTER.FRAP.NUMBER   : num 4
 $ TOT0.SIGMA.FACTOR      : num 3
 $ GEOMVFRAC.SIGMA.FACTOR : num 3
 $ GEOMVFRAC.REL.SD       : num 0.2
 $ ref.img                : chr "lev*.stk.jpg"
 $ time.series            : chr "lev*-time-*.xls"
 $ fname                  : chr "lev*-tif-fname.txt"
 $ desc                   : chr "levData.txt"
 $ OIF.date               : chr "OIF-date.txt"
 $ channel.name           : chr "YFP"
}}
\usage{
param.FRAP.pairs
}
\description{
List of options for the FRAP experiment analysis
}
\details{
this list contains all the options that have to be defined
for the analysis \tabular{ll}{ \code{MODEL.PARAMETERS} \tab
character vector with the model's parameter names, in the
order expected by the matlab functions \cr
\code{FIXED.PARAMETERS} \tab named integer vector, with a 1
for fixed parameters and a 0 for free parameters. Same
order as MODEL.PARAMETERS\cr \code{VALUE.PARAMETERS} \tab
named numeric vector, with initial or fixed values for the
parameters. If 'NA', the parameter is estimated from the
data by \code{\link{calcInitBound}}.\cr
\code{PRE.ACTIVATION.FRAMES} \tab number of frames before
the photo bleaching \cr \code{TOTAL.FRAMES.PER.FRAP} \tab
total number of frames in a FRAP, this includes the pre
activation frames \cr \code{AUTO.F} \tab fix
autofluorescence to substract to the signal. Normally zero
as the auto fluorescence is a parameter of the model.\cr
\code{F.COR.CYT} \tab cytoplasmatic correction factor. A
fix number by which to multiply the cytoplasmic
fluorescence to compensate differences in volumen between
the nucleous and cytoplasm. Normally fixed at one since the
model's parameter geomVfrac takes care of this.\cr
\code{SEC.PER.FRAME} \tab seconds per frame, used to
calculate the time of each frame. This can change between
experiments\cr \code{PB.CORRECTION.FUN} \tab a function,
with arguments fl (fluorescence) and time. Returns the
value of fluorescence corrected by imaging photobleaching.
If photobleaching is neglectable, can be set to the
'identity' function.\cr \code{SIGMA.FUN} \tab a function of
the fluorescnece level that returns the sigma, i.e. the
expected noise level for each fluorescence level. Only used
if SIGMA.METHOD='SIGMA.FUN'\cr \code{SIGMA.METHOD} \tab
string indicating wich method to use to estimate the noise
(sigma). Available options are 'SIGMA.FUN' and
'SINGLE.CELL.PAIR'.  See
\code{\link{getCrossResidualsPair}} for more details on
this last method.\cr \code{FRAP.ORDER.SEP} \tab separator
string for the 'pb' column of the description file
specified by 'desc' (see below)\cr
\code{FRAP.ORDER.DAUGHTER.STR} \tab daughter token for the
pb column of the description file\cr
\code{FRAP.ORDER.MOTHER.STR} \tab mother token for the pb
column of the description file\cr \code{MOTHER.FRAP.NUMBER}
\tab number of FRAPs in the train of FRAPs expected for the
mother\cr \code{DAUGHTER.FRAP.NUMBER} \tab number of FRAPs
in the train of FRAPs expected for the daughter\cr
\code{TOT0.SIGMA.FACTOR} \tab number of standard deviations
to let the 'tot0' parameter wonder about during the fit.
This region is centered in the initial estimation for
'tot0' done by \code{\link{calcInitBound}}\cr
\code{ref.img} \tab string with pattern for the reference
images. That is the images with the nuclear and citoplasmic
regions drawn on it.\cr \code{time.series} \tab string with
pattern for the time-series files. Those are the files
generated by ImageJ's plugin 'Time Series Analyzer v2'\cr
\code{fname} \tab string with pattern for the text file
containing the file-names of all the tif images in the
experiment.\cr \code{desc} \tab string with pattern for the
description file of the experiment. This tab delimited file
should contain the columns 'pos' (or 'lev'), 'pb' and other
description columns\cr \code{OIF.date} \tab string with the
name of the OIF-date file, i.e. the file with all
acuisition dates of the images\cr \code{channel.name} \tab
string with name of the relevant channel\cr }
}
\keyword{datasets}

