\name{getNumber}
\alias{getNumber}
\title{Gets the number on a string}
\usage{
getNumber(fname, after = "")
}
\arguments{
  \item{fname}{string from which to extract numbers}

  \item{after}{string pattern after which the required
  number is located}
}
\value{
a numeric vector, extracted from \code{fname}
}
\description{
Gets the number on a string
}

