\name{gg_color_hue}
\alias{gg_color_hue}
\title{The colors used by dafualt by gglot2 for factors}
\usage{
gg_color_hue(n)
}
\arguments{
  \item{n}{number of colors to return}
}
\value{
string description of colors
}
\description{
The colors used by dafualt by gglot2 for factors
}
\examples{
gg_color_hue(4)
}

