\docType{package}
\name{NuclearFRAP}
\alias{NuclearFRAP}
\alias{NuclearFRAP-package}
\title{Nuclear FRAP Data Analysis Package}
\description{
This package prvides functions to load, visualize and do
model fitting on FRAP data. It also provides functions to
do profile likelihood analysis.
}
\details{
To install NuclearFRAP read \code{vignette("Install")}

To get started with NuclearFRAP chek out the introduction
\code{vignette("NuclearFRAP")}

To see the model defintion and parameters
\code{vignette("Model")}

To see a tutorial on image acquisition and manual analysis
(in spanish) \code{vignette("Tutorial")}
}
\author{
Alan Bush \email{abush@fbmc.fcen.uba.ar}
}

