\name{plot.frapPair}
\alias{plot.frapPair}
\title{plot a FRAP pair experiment}
\usage{
\method{plot}{frapPair}(frapPair, swapParam = TRUE)
}
\arguments{
  \item{frapPair}{a frapPair object, as returned by
  \code{\link{analyzeFrapPair}}}

  \item{swapParam}{boolean indicating if swap fits
  parameters should be included in the parameters table}
}
\description{
plot a FRAP pair experiment
}

