\name{plotBestUpperLowerFrap}
\alias{plotBestUpperLowerFrap}
\title{plot a FRAP experiment, with best lower and upper estimations for a parameter}
\usage{
plotBestUpperLowerFrap(frap, upper, lower)
}
\arguments{
  \item{frap}{a frap object, as returned by
  \code{\link{analyzeFrap}}, with best estimation of
  parameters}

  \item{upper}{a frap object, as returned by
  \code{\link{analyzeFrap}}, with upper boundary for a
  parameter}

  \item{lower}{a frap object, as returned by
  \code{\link{analyzeFrap}}, with lower boundary for a
  parameter}
}
\description{
plot a FRAP experiment, with best lower and upper
estimations for a parameter
}

