\name{plotBestUpperLowerFrapPair}
\alias{plotBestUpperLowerFrapPair}
\title{plot a FRAP pair experiment, with best lower and upper estimations for a parameter}
\usage{
plotBestUpperLowerFrapPair(frapPair, upper, lower)
}
\arguments{
  \item{frap}{a frapPair object, as returned by
  \code{\link{analyzeFrapPair}}, with best estimation of
  parameters}

  \item{upper}{a frapPair object, as returned by
  \code{\link{analyzeFrapPair}}, with upper boundary for a
  parameter}

  \item{lower}{a frapPair object, as returned by
  \code{\link{analyzeFrapPair}}, with lower boundary for a
  parameter}
}
\description{
plot a FRAP pair experiment, with best lower and upper
estimations for a parameter
}

