\name{plotExperimentSummary}
\alias{plotExperimentSummary}
\title{create a scatter plot of kI and kEVfrac for all pairs}
\usage{
plotExperimentSummary(experiment, main = "")
}
\arguments{
  \item{frapPairList}{a list containing 'frapPair' objects,
  as returned by \code{\link{analyzeFrapPair}}}
}
\description{
create a scatter plot of kI and kEVfrac for all pairs
}

