\name{plotResidualsAllFits}
\alias{plotResidualsAllFits}
\title{plot all the fits and respective residuals}
\usage{
plotResidualsAllFits(frap)
}
\arguments{
  \item{frap}{a frap or frapPair object, as returned by
  \code{\link{analyzeFrap}} or
  \code{\link{analyzeFrapPair}}}
}
\description{
plot all the fits and respective residuals
}

