\name{restructureSimData}
\alias{restructureSimData}
\title{restrucutres the simulated data for a single cell}
\usage{
restructureSimData(db)
}
\arguments{
  \item{db}{a data.frame as returned by \code{getSimData}}
}
\value{
a restructured data.frame
}
\description{
restrucutres the simulated data for a single cell
}
\details{
returns a restructured data.frame more suitable for
plotting
}

