\name{restructureSimDataPair}
\alias{restructureSimDataPair}
\title{restrucutres the simulated data for a cell pair}
\usage{
restructureSimDataPair(db)
}
\arguments{
  \item{db}{a data.frame as returned by
  \code{getSimDataPair}}
}
\value{
a restructured data.frame
}
\description{
restrucutres the simulated data for a cell pair
}
\details{
returns a restructured data.frame more suitable for
plotting
}

