\name{runProfileLikelihoodContourPair}
\alias{runProfileLikelihoodContourPair}
\title{run the profile likelihhod analysis on 'fix directions' from the fitted point, for a pair of cells}
\usage{
runProfileLikelihoodContourPair(...)
}
\value{
a data.frame with the points along the directions defined
by theta where cost = min.cost + boundaryCost
}
\description{
run the profile likelihhod analysis on 'fix directions'
from the fitted point, for a pair of cells
}

