\name{runProfileLikelihoodCostPair}
\alias{runProfileLikelihoodCostPair}
\title{get the profile likelihood cost for a point in the parameter space for a cell pair}
\usage{
runProfileLikelihoodCostPair(...)
}
\arguments{
  \item{...}{pair arguments for runProfileLikelihoodCost}
}
\value{
a data.frame with the point in the parameter space and the
profile likelihood cost, and a 'type' column
}
\description{
get the profile likelihood cost for a point in the
parameter space for a cell pair
}
\details{
pair version of \code{runProfileLikelihoodCost}
}

