\name{runProfileLikelihoodGridPair}
\alias{runProfileLikelihoodGridPair}
\title{run the profile likelihhod analysis on a regular grid of points in the parameter space, for a cell pair}
\usage{
runProfileLikelihoodGridPair(...)
}
\value{
a data.frame with the points in the parameter space and the
profile likelihood cost, for the cell pair
}
\description{
run the profile likelihhod analysis on a regular grid of
points in the parameter space, for a cell pair
}
\details{
pair version of \code{\link{runProfileLikelihoodGrid}}. db,
param, and expData should contain info about both cells
}

