\name{transformTauNucFrac}
\alias{transformTauNucFrac}
\title{Calculate tau and nucFrap from kI and kEVfrac}
\usage{
transformTauNucFrac(x)
}
\arguments{
  \item{x}{a frap, frapPair or experimentSummary object}
}
\value{
input object with modified param and PL.db alements
}
\description{
Calculate tau and nucFrap from kI and kEVfrac
}

